+++ src/engine/pkt_terminal_internal.h
@@ -10,8 +10,8 @@ int __pkt_terminal_update(void);
 int __pkt_terminal_init(struct pkt_config *config);
 int __pkt_terminal_restore(void);
 int __pkt_terminal_read_input(void);
-int __pkt_terminal_putc(int x, int y, enum pkt_color fcolor, enum pkt_color bcolor, uint8_t attr, char c);
-int __pkt_terminal_puts(int x, int y, enum pkt_color fcolor, enum pkt_color bcolor, uint8_t attr, const char *str);
+int __pkt_terminal_putc(int x, int y, enum pkt_color fcolor, enum pkt_color bcolor, unsigned int attr, char c);
+int __pkt_terminal_puts(int x, int y, enum pkt_color fcolor, enum pkt_color bcolor, unsigned int attr, const char *str);
 
 /*
  * Check terminal resize and store current cols and rows to the pointer passed as parameters.
